Add message content
