+++ generator/daemon/__main__.py
@@ -1,9 +1,11 @@
+import json
 import time
 from signal import SIGINT, signal
 
 from aws import s3, sqs
 
 from . import POLL_INTERVAL_SECONDS, SQS_INPUT_QUEUE
+from .. import pdf
 
 
 def shutdown(signal, frame):
@@ -14,10 +16,16 @@ def shutdown(signal, frame):
 signal(SIGINT, shutdown)
 
 
+def maker(message):
+    content = json.loads(message.body['document']['content'])
+
+    print(f"Making ({type(content)}) {content}")
+
+
 def logger(message):
     print(message, flush=True)
 
 
 sqs.poller.start(queue=SQS_INPUT_QUEUE,
                  poll_interval=POLL_INTERVAL_SECONDS,
-                 handlers=[logger])
+                 handlers=[maker, logger])
